Tweaks

+++ census.sh
@@ -28,12 +28,17 @@ done
 for repo in $(find $org -mindepth 1 -maxdepth 1 -type d); do
   if [ -e $repo/.scan ] && [ ! -s $repo/.scan ]; then
     sha=$(curl -sSL "https://api.github.com/repos/$repo/branches/$branch" | jq -r '.commit.sha // empty')
-    echo "$sha" > "$repo/.scan"
+    last_sha=$(cat "$repo/.scan" || true)
     if [ -n "$sha" ]; then
-      mkdir -p "$repo/$config"
-      for file in $(echo "$files"); do
-        curl -fsL "https://raw.githubusercontent.com/$repo/$sha/$config/$file" -o "$repo/$config/$file" || touch "$repo/$config/$file"
-      done
+      if [ "$sha" != "$last_sha" ]; then
+        echo "$sha" > "$repo/.scan"
+        mkdir -p "$repo/$config"
+        for file in $(echo "$files"); do
+          curl -fsL "https://raw.githubusercontent.com/$repo/$sha/$config/$file" -o "$repo/$config/$file" || touch "$repo/$config/$file"
+        done
+      fi
+    elif [ ! -e "$repo/.scan" ]; then
+      echo "$sha" > "$repo/.scan"
     fi
   fi
 done
